+++ CHANGES.rst
@@ -15,6 +15,7 @@ Unreleased
 - Migrate from the unmaintained arrow2 arrow implementation to the official arrow-rs from apache. This comes along
   with a few changes:
   - `directededges_to_wkb_lines` and `directededges_to_lines` have been removed. Use the linestring-versions instead.
+  - Geometry collections are currently unsupported when working with WKB. This is still work-in-progress within geoarrow-rs.
 
 0.19.2 - 2023-11-16
 -------------------